Refactoring: Extract bidAdjustments code from exchangeService
