+++ src/main/java/org/prebid/server/util/ListUtil.java
@@ -1,5 +1,6 @@
 package org.prebid.server.util;
 
+import org.apache.commons.collections4.CollectionUtils;
 import org.prebid.server.util.algorithms.ListsUnionView;
 
 import java.util.List;
@@ -11,5 +12,9 @@ public class ListUtil {
 
     public static <T> List<T> union(List<? extends T> first, List<? extends T> second) {
         return new ListsUnionView<>(first, second);
+    }
+
+    public static <T> List<T> nullIfEmpty(List<T> value) {
+        return CollectionUtils.isEmpty(value) ? null : value;
     }
 }
